refactor: refactor the logger module
