+++ src/common/logger/index.ts
@@ -1,0 +1,3 @@
+import { logger } from "./logger";
+
+export { logger }
